services/harvester/requirements.txt
httpx==0.27.2
tenacity==9.0.0
pydantic==2.9.2
python-dotenv==1.0.1
